% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ca_fips.R
\name{calculate_acs_metrics}
\alias{calculate_acs_metrics}
\title{Calculate derived ACS metrics}
\usage{
calculate_acs_metrics(acs_raw)
}
\arguments{
\item{acs_raw}{Raw ACS data from fetch_acs_data}
}
\value{
List with calculated metrics
}
\description{
Calculate derived ACS metrics
}
\keyword{internal}
