% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{current}
\alias{current}
\title{Analyze current year chronic absence patterns}
\usage{
current(data, current_year)
}
\arguments{
\item{data}{Student-level data frame}

\item{current_year}{Numeric current year}
}
\value{
List containing current year analysis
}
\description{
Analyze current year chronic absence patterns
}
\keyword{internal}
