% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Student engagement data for California districts}
\format{
A data frame with 738,712 observations and 15 variables:
\describe{
\item{COUNTY_ID}{County identifier}
\item{DISTRICT_ID}{District identifier}
\item{SCHOOL_ID}{School identifier}
\item{COUNTY_NAME}{County name}
\item{DISTRICT_NAME}{District name}
\item{SCHOOL_NAME}{School name}
\item{DISAGGREGATED_GROUP}{Student demographic group}
\item{YEAR}{Academic year (2018-2023)}
\item{CHRONIC_ABSENT_ELIGIBLE}{Number of students eligible for chronic absenteeism calculation}
\item{CHRONIC_ABSENT_COUNT}{Number of chronically absent students}
\item{CHRONIC_ABSENT_RATE}{Chronic absenteeism rate (percentage)}
\item{STUDENTS_ENROLLED}{Number of students enrolled}
\item{STUDENTS_TESTED}{Number of students tested}
\item{SUBJECT}{Assessment subject (ELA or MATH)}
\item{PARTICIPATION_RATE}{Assessment participation rate (percentage)}
}
}
\source{
California Department of Education
}
\usage{
df
}
\description{
A dataset containing assessment participation and chronic absenteeism
data for California school districts.
}
\keyword{datasets}
