% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{district_county_context}
\alias{district_county_context}
\title{Analyze district in county context}
\usage{
district_county_context(district_data, county_name, acs_year)
}
\arguments{
\item{district_data}{Filtered district data}

\item{county_name}{County name}

\item{acs_year}{ACS year}
}
\value{
List with county context analysis
}
\description{
Analyze district in county context
}
\keyword{internal}
