% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{district_grades}
\alias{district_grades}
\title{District grade analysis}
\usage{
district_grades(district_data, current_year)
}
\arguments{
\item{district_data}{Filtered district data}

\item{current_year}{Current year}
}
\value{
List with grade-level analysis
}
\description{
District grade analysis
}
\keyword{internal}
