% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ca_fips.R
\name{fetch_acs_data}
\alias{fetch_acs_data}
\title{Fetch ACS data for a county}
\usage{
fetch_acs_data(state_fips, county_fips, year = 2022)
}
\arguments{
\item{state_fips}{State FIPS code}

\item{county_fips}{County FIPS code}

\item{year}{ACS year (default: 2022)}
}
\value{
Data frame with ACS variables
}
\description{
Fetch ACS data for a county
}
\keyword{internal}
