% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ca_fips.R
\name{get_county_acs_data}
\alias{get_county_acs_data}
\title{Get ACS data for a county by name}
\usage{
get_county_acs_data(county_name, year = 2022)
}
\arguments{
\item{county_name}{County name}

\item{year}{ACS year (default: 2022)}
}
\value{
List with ACS metrics
}
\description{
Get ACS data for a county by name
}
\keyword{internal}
