% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_means.R
\name{group_means}
\alias{group_means}
\title{Calculate Mean Chronic Absenteeism Rates by Subgroup and Year}
\usage{
group_means(data)
}
\arguments{
\item{data}{Student-level data frame with demographic columns}
}
\value{
Data frame with mean rates by year and subgroup
}
\description{
Calculates state-level mean chronic absenteeism
rates for each student subgroup.
}
\keyword{internal}
