% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandemic.R
\name{pandemic}
\alias{pandemic}
\title{Calculate Pre-Pandemic Comparison (2018 vs 2023)}
\usage{
pandemic(state_data, county_data, district_data, school_data)
}
\arguments{
\item{state_data}{Results from group_means()}

\item{county_data}{Results from county_means()}

\item{district_data}{Results from district_means()}

\item{school_data}{Results from school_means()}
}
\value{
List containing pre-pandemic comparisons for all levels
}
\description{
Calculates percentage point changes and percentage changes from
pre-pandemic (2018) to current (2023) levels for all geographic levels.
}
\keyword{internal}
