% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_district.R
\name{render_district}
\alias{render_district}
\title{Generate district chronic absence report}
\usage{
render_district(data_file, district_name)
}
\arguments{
\item{data_file}{Path to student-level chronic absence data file}

\item{district_name}{Name of district to analyze}
}
\description{
Generate district chronic absence report
}
