% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_county_context.R
\name{text_county_context}
\alias{text_county_context}
\title{Generate county context analysis text}
\usage{
text_county_context(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with county context analysis text
}
\description{
Generate county context analysis text
}
\keyword{internal}
