% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_demographics.R
\name{text_demographics}
\alias{text_demographics}
\title{Generate demographic analysis text}
\usage{
text_demographics(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with demographic analysis text
}
\description{
Generate demographic analysis text
}
\keyword{internal}
