% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_executive.R
\name{text_executive}
\alias{text_executive}
\title{Generate executive summary text}
\usage{
text_executive(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with executive summary text
}
\description{
Generate executive summary text
}
\keyword{internal}
