% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_grades.R
\name{text_grades}
\alias{text_grades}
\title{Generate grade analysis text}
\usage{
text_grades(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with grade analysis text
}
\description{
Generate grade analysis text
}
\keyword{internal}
