% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_metrics.R
\name{text_metrics}
\alias{text_metrics}
\title{Generate district metrics text}
\usage{
text_metrics(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with district metrics text
}
\description{
Generate district metrics text
}
\keyword{internal}
