% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_overview.R
\name{text_overview}
\alias{text_overview}
\title{Generate district overview text}
\usage{
text_overview(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with district overview text
}
\description{
Generate district overview text
}
\keyword{internal}
