% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{district_demographics}
\alias{district_demographics}
\title{District demographic analysis}
\usage{
district_demographics(district_data, current_year)
}
\arguments{
\item{district_data}{Filtered district data}

\item{current_year}{Current year}
}
\value{
List with demographic analysis
}
\description{
District demographic analysis
}
