% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{district_overview}
\alias{district_overview}
\title{District overview analysis}
\usage{
district_overview(district_data, current_year)
}
\arguments{
\item{district_data}{Filtered district data}

\item{current_year}{Current year}
}
\value{
List with overview metrics
}
\description{
District overview analysis
}
