% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{district_schools}
\alias{district_schools}
\title{District school analysis}
\usage{
district_schools(district_data, current_year)
}
\arguments{
\item{district_data}{Filtered district data}

\item{current_year}{Current year}
}
\value{
List with school-level analysis
}
\description{
District school analysis
}
