% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_benchmarks.R
\name{text_benchmarks}
\alias{text_benchmarks}
\title{Generate benchmark analysis text}
\usage{
text_benchmarks(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with benchmark analysis text
}
\description{
Generate benchmark analysis text
}
\keyword{internal}
