% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_schools.R
\name{text_schools}
\alias{text_schools}
\title{Generate school analysis text}
\usage{
text_schools(analysis_results)
}
\arguments{
\item{analysis_results}{Results from analyze_district function}
}
\value{
Character string with school analysis text
}
\description{
Generate school analysis text
}
\keyword{internal}
